+++ test/functional/abc_rpc_isfinal.py
@@ -6,10 +6,21 @@
 import random
 
 from test_framework.address import ADDRESS_ECREG_UNSPENDABLE
+from test_framework.authproxy import JSONRPCException
 from test_framework.avatools import AvaP2PInterface
-from test_framework.messages import AvalancheVote, AvalancheVoteError
+from test_framework.blocktools import create_block, create_coinbase
+from test_framework.messages import (
+    AvalancheVote,
+    AvalancheVoteError,
+    CBlockHeader,
+    msg_headers,
+)
 from test_framework.test_framework import BitcoinTestFramework
-from test_framework.util import assert_raises_rpc_error, uint256_hex
+from test_framework.util import (
+    assert_equal,
+    assert_raises_rpc_error,
+    uint256_hex,
+)
 
 QUORUM_NODE_COUNT = 16
 
@@ -31,6 +42,22 @@ class AvalancheIsFinalTest(BitcoinTestFramework):
 
     def run_test(self):
         node = self.nodes[0]
+
+        tip = node.getbestblockhash()
+
+        assert_raises_rpc_error(
+            -1,
+            "Avalanche is not ready to poll yet.",
+            self.nodes[0].isfinalblock,
+            tip,
+        )
+        assert_raises_rpc_error(
+            -1,
+            "Avalanche is not ready to poll yet.",
+            self.nodes[0].isfinaltransaction,
+            node.getblock(tip)['tx'][0],
+            tip,
+        )
 
         # Build a fake quorum of nodes.
         def get_quorum():
@@ -95,12 +122,17 @@ class AvalancheIsFinalTest(BitcoinTestFramework):
         if self.is_wallet_compiled():
             self.log.info("Check mempool transactions are not finalized")
             # Mature some utxos
-            node.generate(100)
+            tip = node.generate(100)[-1]
             wallet_txid = node.sendtoaddress(
                 ADDRESS_ECREG_UNSPENDABLE, 1_000_000)
             assert wallet_txid in node.getrawmempool()
-            assert not node.isfinaltransaction(
-                wallet_txid, node.getbestblockhash())
+            assert_raises_rpc_error(
+                -5,
+                "No such transaction found in the provided block.",
+                node.isfinaltransaction,
+                wallet_txid,
+                tip,
+            )
 
             self.log.info(
                 "A transaction is only finalized if the containing block is finalized")
@@ -109,15 +141,43 @@ class AvalancheIsFinalTest(BitcoinTestFramework):
             assert not node.isfinaltransaction(wallet_txid, tip)
             self.wait_until(lambda: is_finalblock(tip))
             assert node.isfinaltransaction(wallet_txid, tip)
+
             # Needs -txindex
-            assert not node.isfinaltransaction(wallet_txid)
+            assert_raises_rpc_error(
+                -5,
+                "No such transaction. Use -txindex or provide a block hash to enable blockchain transaction queries.",
+                node.isfinaltransaction,
+                wallet_txid,
+            )
 
             self.log.info(
                 "Repeat with -txindex so we don't need the blockhash")
             self.restart_node(0, self.extra_args[0] + ['-txindex'])
-
             quorum = get_quorum()
             self.wait_until(is_quorum_established)
+
+            # Try to raise a -txindex not synced yet error. This is not
+            # guaranteed because syncing is fast!
+            try:
+                node.isfinaltransaction(
+                    uint256_hex(random.randint(0, 2**256 - 1)),
+                )
+            except JSONRPCException as e:
+                assert_equal(e.error['code'], -5)
+
+                if e.error['message'] == "No such mempool or blockchain transaction.":
+                    # If we got a regular  "not found" error, the txindex should
+                    # have synced.
+                    assert node.getindexinfo()['txindex']['synced'] is True
+                else:
+                    # Otherwise we might have successfully raised before the
+                    # indexer completed. Checking the status now is useless as
+                    # the indexer might have completed the synchronization in
+                    # the meantime and the status is no longer relevant.
+                    assert e.error['message'] == "No such transaction. Blockchain transactions are still in the process of being indexed."
+            else:
+                assert False, "The isfinaltransaction RPC call did not throw as expected."
+
             self.wait_until(lambda: node.getindexinfo()[
                             'txindex']['synced'] is True)
 
@@ -129,8 +189,12 @@ class AvalancheIsFinalTest(BitcoinTestFramework):
             assert wallet_txid in node.getrawmempool()
             assert not node.isfinaltransaction(wallet_txid)
 
-            assert not node.isfinaltransaction(
-                uint256_hex(random.randint(0, 2**256 - 1)))
+            assert_raises_rpc_error(
+                -5,
+                "No such mempool or blockchain transaction.",
+                node.isfinaltransaction,
+                uint256_hex(random.randint(0, 2**256 - 1)),
+            )
 
         self.log.info("Check unknown item")
         for _ in range(10):
@@ -147,6 +211,26 @@ class AvalancheIsFinalTest(BitcoinTestFramework):
                 uint256_hex(random.randint(0, 2**256 - 1)),
                 uint256_hex(random.randint(0, 2**256 - 1)),
             )
+
+        tip = node.getbestblockhash()
+        height = node.getblockcount() + 1
+        time = node.getblock(tip)['time'] + 1
+        block = create_block(int(tip, 16), create_coinbase(height), time)
+        block.solve()
+
+        peer = node.add_p2p_connection(AvaP2PInterface())
+        msg = msg_headers()
+        msg.headers = [CBlockHeader(block)]
+        peer.send_message(msg)
+
+        self.wait_until(lambda: node.getchaintips()[0]['height'] == height)
+        assert_raises_rpc_error(
+            -1,
+            "Block data not downloaded yet.",
+            node.isfinaltransaction,
+            uint256_hex(random.randint(0, 2**256 - 1)),
+            uint256_hex(block.sha256),
+        )
 
 
 if __name__ == '__main__':