+++ doc/release-notes.md
@@ -8,3 +8,5 @@ This release includes the following features and fixes:
  - The deprecated `addnode` and `whitelisted` have been removed from the
    `getpeerinfo` RPC output. Users relying on the `deprecatedrpc` behavior need
    to update their system accordingly.
+ - The `isfinaltransaction` RPC now returns a detailed error when the
+   transaction is not found, instead of simply returning false.